refactor(audio): 重构音频设备类型和接口设计
- 将 AudioDeviceType 枚举从 int 类型改为 char 类型以节省内存
- 添加 AAudio 音频后端支持并重新排列设备类型枚举值
- 为所有音频相关的常量和函数添加 [[nodiscard]] 属性
- 移除 CMakeLists.txt 中的 Android 平台 vendor 库配置
- 简化 CMakeLists.txt 中的源文件收集逻辑并移除硬编码的音频后端注册
- 修复 Windows 平台输入驱动中的 RegisterRawInputDevices 函数声明和实现
- 修改 PlatformWindows.cpp 中鼠标按键状态检查的类型转换逻辑

+++ src/engine/PlatformWindows.cpp
@@ -351,7 +351,7 @@ bool Platform::IsKeyDown(PrismaEngine::Input::KeyCode key) {
 }
 
 bool Platform::IsMouseButtonDown(PrismaEngine::Input::MouseButton btn) {
-    SHORT state = GetAsyncKeyState(btn);
+    SHORT state = GetAsyncKeyState(static_cast<int>(btn));
     return (state & 0x8000) != 0;
 }
